Refactors File Structure
