--- src/App.js
@@ -1,17 +1,0 @@
-
-import React, { Component } from "react";
-import Login from './login'
-import Main from './main'
- 
-class App extends Component {
-  render() {
-    return (
-      <div>
-        {/* <Login /> */}
-        <Main />
-      </div>
-    );
-  }
-}
- 
-export default App;
