+++ src/Components/MainPage/MainPage.css
@@ -1,0 +1,41 @@
+.container {
+    display: flex;
+    flex-direction: row;
+    height: 100vh;
+    align-items: stretch;
+    justify-content: space-between;
+    background-color:#A8C3BC;
+    
+  }
+  
+  .mainLeft {
+    border-style: solid;
+    flex: 1 1 20%;
+    display: flex;
+  }
+  
+  .mainCenter {
+    border-style: solid;
+    flex: 1 1 60%;
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .mainRight {
+    border-style: solid;
+    flex: 1 1 20%;
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .mainRight > * {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  img {
+    width: 60%;
+    height: auto;
+    display: block;
+  }
