+++ src/Components/Menu/Menu.css
@@ -1,0 +1,6 @@
+.Menu {
+    flex: 1 1 auto;
+    border: solid;
+    border-width: 1px;
+    text-align: center;
+  }
