+++ src/Components/Menu/Menu.js
@@ -1,0 +1,7 @@
+import "./Menu.css"
+
+function Menu () {
+    return <div className="Menu ">Menu</div>
+}
+
+export default Menu
