+++ src/Components/ProgressBar/ProgressBar.js
@@ -1,0 +1,12 @@
+import "./ProgressBar.css"
+
+//https://css-tricks.com/css3-progress-bars/
+function ProgressBar () {    
+    return (
+    <div class="meter orange nostripes">
+        
+        </div>
+    )
+}
+
+export default ProgressBar
