+++ src/Components/Timer/Timer.css
@@ -1,0 +1,7 @@
+.Timer {
+    flex: 1 1 50%;
+    border: solid;
+    border-width: 1px;
+    text-align: center;
+  }
+
