+++ src/Components/Timer/Timer.js
@@ -1,0 +1,10 @@
+import "./Timer.css"
+
+function Timer () {  
+    return (
+    <div className="Timer">
+        <img src="http://clipart-library.com/newimages/clip-art-clock-14.png" alt="tree"></img>
+    </div>)
+}
+
+export default Timer
