+++ src/Components/Tree/Tree.js
@@ -1,0 +1,11 @@
+import "./Tree.css"
+
+function Tree () {   
+    return (
+    <div className="Tree">
+        <img src='https://cdn.pixabay.com/photo/2014/12/22/00/07/tree-576847__480.png' alt="tree"></img> 
+    </div>
+    )
+}
+
+export default Tree
